+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added docx exporters for space reports in myems-api and myems-web
+- added pdf exporters for equipment reports in myems-api and myems-web
 ### Changed
 ### Fixed
 ### Removed